reorganized files to match membership_portal branch folder structure
